Move Swapchain files into compatiblity folder
First step in preparing to remove swapchain behavior from being
hard-coded inside the decode framework so it can be used to
build a compatibility library for use in additional scenarios.
