+++ CMakeLists.txt
@@ -305,6 +305,7 @@ if (${RUN_TESTS})
 endif()
 
 add_subdirectory(external/nlohmann)
+add_subdirectory(compatibility)
 add_subdirectory(framework)
 if(NOT BUILD_STATIC)
     add_subdirectory(layer)
