+++ framework/decode/vulkan_replay_consumer_base.cpp
@@ -25,8 +25,8 @@
 #include "decode/custom_vulkan_struct_handle_mappers.h"
 #include "decode/descriptor_update_template_decoder.h"
 #include "decode/resource_util.h"
-#include "decode/vulkan_captured_swapchain.h"
-#include "decode/vulkan_virtual_swapchain.h"
+#include "compatibility/vulkan_captured_swapchain.h"
+#include "compatibility/vulkan_virtual_swapchain.h"
 #include "decode/vulkan_enum_util.h"
 #include "decode/vulkan_feature_util.h"
 #include "decode/vulkan_object_cleanup_util.h"
