+++ tools/extract/CMakeLists.txt
@@ -46,7 +46,7 @@ endif()
 
 target_include_directories(gfxrecon-extract PUBLIC ${CMAKE_BINARY_DIR})
 
-target_link_libraries(gfxrecon-extract gfxrecon_decode gfxrecon_graphics gfxrecon_format gfxrecon_util platform_specific)
+target_link_libraries(gfxrecon-extract gfxrecon_compatibility gfxrecon_decode gfxrecon_graphics gfxrecon_format gfxrecon_util platform_specific)
 
 common_build_directives(gfxrecon-extract)
