+++ tools/optimize/CMakeLists.txt
@@ -58,6 +58,7 @@ target_include_directories(gfxrecon-optimize PUBLIC ${CMAKE_BINARY_DIR})
 
 target_link_libraries(gfxrecon-optimize
                           gfxrecon_application
+                          gfxrecon_compatibility
                           gfxrecon_decode
                           gfxrecon_graphics
                           gfxrecon_format
